fix: Extract parseRssItems into its own file
Move the parseRssItems implementation out of lib/index.ts into a new lib/parseRssItems.ts and update lib/index.ts to re-export parseRssItems, goodreads, and substack. Also simplify the Vite build entry to use the single barrel ./lib/index.ts instead of listing multiple entry files, so the library is built from the unified exports.
